framepath - workaround for AIMExt 19.10.12 number formatting bug
'aimext' executable of AIMAll 19.10.12 sometimes overwrites the 'E' in a floating point number string ending with, e.g. 'E-02' if the exponent has 3 digits, giving a string of the form '-dddddddddd-ddd' instead of  '-ddddddddddE-ddd', choking the Python3 float() conversion function. Workaround checks for the presence of a '+' or '-' in the fourth-to-last character and re-inserts the 'E' if necessary.